long value shouldn't be parsed as double

+++ twitter4j-core/src/internal-json/java/twitter4j/JSON.java
@@ -75,7 +75,7 @@ class JSON {
             return ((Number) value).longValue();
         } else if (value instanceof String) {
             try {
-                return (long) Double.parseDouble((String) value);
+                return Long.parseLong((String) value);
             } catch (NumberFormatException ignored) {
             }
         }
